generic component of in progress
